# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "FKMed",
    "description" : "FKMed — health plan beneficiary portal API",
    "version" : "0.12.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/sim/tele/sessions/{id}/close" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "close",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CloseTeleSessionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimTeleSessionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/tele/sessions/next/start" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "startNext",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartNextTeleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimTeleSessionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/reimbursements/{id}/pendency" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "openReimbursementPendency",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimOpenReimbursementPendencyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimReimbursementResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/reimbursements/{id}/pay" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "payReimbursement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimPayReimbursementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimReimbursementResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/reimbursements/{id}/deny" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "denyReimbursement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimDenyReimbursementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimReimbursementResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/reimbursements/{id}/approve" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "approveReimbursement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimReimbursementResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/reimbursement-previews/{id}/conclude" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "concludePreview",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimConcludePreviewRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimPreviewResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/guides" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "createGuide",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimCreateGuideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimGuideResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/guides/{id}/partially-authorize" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "partiallyAuthorizeGuide",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimPartiallyAuthorizeGuideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimGuideResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/guides/{id}/mark-executed" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "markGuideExecuted",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimGuideResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/guides/{id}/deny" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "denyGuide",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimDenyGuideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimGuideResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/guides/{id}/cancel" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "cancelGuide",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimGuideResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/guides/{id}/authorize" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "authorizeGuide",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimAuthorizeGuideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimGuideResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/finance/invoices" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "createInvoice",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimCreateInvoiceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimInvoiceResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/finance/invoices/{id}/pay" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "payInvoice",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/sim/finance/copay" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "createCopay",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SimCreateCopayRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimCopayResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/sim/documents" : {
      "post" : {
        "tags" : [ "operator-simulation" ],
        "operationId" : "issueDocument",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IssueOperatorDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimIssuedDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/recovery/request" : {
      "post" : {
        "tags" : [ "recovery-controller" ],
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordRecoveryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        }
      }
    },
...
    "/api/tokens/current" : {
      "get" : {
        "tags" : [ "token-controller" ],
        "operationId" : "current",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/network/providers/{id}" : {
      "get" : {
        "tags" : [ "network-controller" ],
        "operationId" : "provider",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProviderDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/legal-documents/{type}" : {
      "get" : {
        "tags" : [ "legal-document-controller" ],
        "operationId" : "document",
        "parameters" : [ {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LegalDocumentView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/guides" : {
      "get" : {
        "tags" : [ "guide-controller" ],
        "operationId" : "list_3",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "AUTORIZADA", "PARCIALMENTE_AUTORIZADA", "NEGADA", "CANCELADA", "EXECUTADA" ]
          }
        }, {
          "name" : "period",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "LAST_30D", "LAST_90D", "LAST_12M" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/GuideListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/finance/invoices" : {
      "get" : {
        "tags" : [ "finance-controller" ],
        "operationId" : "invoices",
        "parameters" : [ {
          "name" : "tab",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "OPEN",
            "enum" : [ "OPEN", "PAID" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/InvoiceSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/finance/invoices/{id}" : {
      "get" : {
        "tags" : [ "finance-controller" ],
        "operationId" : "invoice",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/finance/copay" : {
      "get" : {
        "tags" : [ "finance-controller" ],
        "operationId" : "copay",
        "parameters" : [ {
          "name" : "period",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "CURRENT_MONTH",
            "enum" : [ "CURRENT_MONTH", "LAST_3M", "CUSTOM" ]
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CopayStatement"
                }
              }
            }
          }
        }
      }
    },
    "/api/context/beneficiaries/{beneficiaryId}" : {
      "get" : {
        "tags" : [ "context-controller" ],
        "operationId" : "beneficiary",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BeneficiarySummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/beneficiaries/{beneficiaryId}/profile" : {
      "get" : {
        "tags" : [ "profile-controller" ],
        "operationId" : "profile",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProfileView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CloseTeleSessionRequest" : {
        "type" : "object",
        "properties" : {
          "professionalName" : {
            "type" : "string",
            "minLength" : 1
          },
          "crm" : {
            "type" : "string",
            "minLength" : 1
          },
          "guidance" : {
            "type" : "string"
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SimDocumentSpec"
            }
          }
        },
        "required" : [ "crm", "professionalName" ]
      },
...
      "SimTeleSessionResult" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "state" : {
            "type" : "string"
          },
          "issuedDocumentIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }
      },
      "StartNextTeleRequest" : {
        "type" : "object",
        "properties" : {
          "professionalName" : {
            "type" : "string",
            "minLength" : 1
          },
          "crm" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "crm", "professionalName" ]
      },
      "SimOpenReimbursementPendencyRequest" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "description" ]
      },
      "SimReimbursementResult" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          },
          "expectedPaymentDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "SimPayReimbursementRequest" : {
        "type" : "object",
        "properties" : {
          "outcome" : {
            "type" : "string",
            "enum" : [ "SUCCESS", "FAILURE" ]
          },
          "failureReason" : {
            "type" : "string"
          }
        }
      },
      "SimDenyReimbursementRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "reason" ]
      },
      "SimConcludePreviewRequest" : {
        "type" : "object",
        "properties" : {
          "estimatedValue" : {
            "type" : "number"
          }
        },
        "required" : [ "estimatedValue" ]
      },
      "SimPreviewResult" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "situation" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "CONCLUIDA" ]
          },
          "estimatedValue" : {
            "type" : "number"
          }
        }
      },
...
      "SimCreateGuideRequest" : {
        "type" : "object",
        "properties" : {
          "beneficiaryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "CONSULTA", "SP_SADT", "INTERNACAO" ]
          },
          "requestingProvider" : {
            "type" : "string",
            "minLength" : 1
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GuideItemRequest"
            },
            "minItems" : 1
          }
        },
        "required" : [ "beneficiaryId", "items", "requestingProvider", "type" ]
      },
      "SimGuideResult" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "number" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "AUTORIZADA", "PARCIALMENTE_AUTORIZADA", "NEGADA", "CANCELADA", "EXECUTADA" ]
          }
        }
      },
...
      "SimPartiallyAuthorizeGuideRequest" : {
        "type" : "object",
        "properties" : {
          "password" : {
            "type" : "string",
            "minLength" : 1
          },
          "validUntil" : {
            "type" : "string",
            "format" : "date"
          },
          "itemStatuses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ItemStatus"
            },
            "minItems" : 1
          }
        },
        "required" : [ "itemStatuses", "password", "validUntil" ]
      },
...
    }
  }
}